# pkg/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  models.Message:
    properties:
      message:
        description: A standard message.
        type: string
    type: object
...
  models.UserGeneralSettings:
    properties:
      default_project_id:
        type: integer
      discoverable_by_email:
        type: boolean
      discoverable_by_name:
        type: boolean
      email_reminders_enabled:
        type: boolean
      extra_settings_links:
        additionalProperties: {}
        description: Server/OpenID-provided; populated on read, ignored on write.
        type: object
      frontend_settings: {}
      language:
        type: string
      name:
        type: string
      overdue_tasks_reminders_enabled:
        type: boolean
      overdue_tasks_reminders_time:
        type: string
      timezone:
        type: string
      week_start:
        maximum: 6
        minimum: 0
        type: integer
    type: object
...
  user.User:
    properties:
      bot_owner_id:
        description: |-
          BotOwnerID is the ID of the owning (human) user if this user is a bot.
          A non-zero value means this user is a bot and cannot authenticate via password.
        type: integer
      created:
        description: A timestamp when this task was created. You cannot change this
          value.
        type: string
      email:
        description: The user's email address.
        maxLength: 250
        type: string
      id:
        description: The unique, numeric id of this user.
        type: integer
      name:
        description: The full name of the user.
        type: string
      updated:
        description: A timestamp when this task was last updated. You cannot change
          this value.
        type: string
      username:
        description: The username of the user. Is always unique.
        maxLength: 250
        minLength: 1
        type: string
    type: object
...
  v1.UserAvatarProvider:
    properties:
      avatar_provider:
        description: The avatar provider. Valid types are `gravatar` (uses the user
          email), `upload`, `initials`, `marble` (generates a random avatar for each
          user), `ldap` (synced from LDAP server), `openid` (synced from OpenID provider),
          `default`.
        type: string
    type: object
...
  web.HTTPError:
    properties:
      code:
        type: integer
      i18n_params:
        additionalProperties:
          type: string
        description: I18nParams carries Message's dynamic values, keyed by the client's
          translation placeholder names, so clients can localise the error.
        type: object
      message:
        type: string
    type: object
info:
  contact:
    email: hello@braznmngo.com
    name: Brazn Tasks contact
    url: https://github.com/BraznMngo/brazn-tasks/issues
  description: |-
    Brazn Tasks is a modified fork of [Vikunja](https://vikunja.io) v2.4.0. Source: https://github.com/BraznMngo/brazn-tasks
    # Pagination
    Every endpoint capable of pagination will return two headers:
    * `x-pagination-total-pages`: The total number of available pages for this request
    * `x-pagination-result-count`: The number of items returned for this request.
    # Permissions
    All endpoints which return a single item (project, task, etc.) - no array - will also return a `x-max-permission` header with the max permission the user has on this item as an int where `0` is `Read Only`, `1` is `Read & Write` and `2` is `Admin`.
    This can be used to show or hide ui elements based on the permissions the user has.
    # Errors
    All errors have an error code and a human-readable error message in addition to the http status code. You should always check for the status code in the response, not only the http status code.
    Due to limitations in the swagger library we're using for this document, only one error per http status code is documented here. Make sure to check the [error docs](https://vikunja.io/docs/errors/) in the upstream Vikunja documentation for a full list of available error codes. Brazn Tasks is a modified fork of Vikunja and the error codes are unchanged.
    # Authorization
...

...

...
paths:
...
  /test/{table}:
    patch:
      consumes:
      - application/json
      description: 'Fills the specified table with the content provided in the payload.
        You need to enable the testing endpoint before doing this and provide the
        `Authorization: <token>` secret when making requests to this endpoint. See
        docs for more details.'
      parameters:
      - description: The table to reset
        in: path
        name: table
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Everything has been imported successfully.
          schema:
            items:
              $ref: '#/definitions/user.User'
            type: array
        "500":
          description: Internal server error.
          schema:
            $ref: '#/definitions/models.Message'
      summary: Reset the db to a defined state
      tags:
      - testing
  /test/all:
    delete:
      description: Removes all data from every Brazn Tasks table. Used by e2e tests to
        ensure clean state before each test. Requires the testing token.
      produces:
      - application/json
      responses:
        "200":
          description: All tables truncated.
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/web.HTTPError'
        "500":
          description: Internal server error.
          schema:
            $ref: '#/definitions/models.Message'
      summary: Truncate all tables
      tags:
      - testing
...
  /user/settings/avatar:
    get:
      consumes:
      - application/json
      description: Returns the current user's avatar setting.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.UserAvatarProvider'
        "400":
          description: Something's invalid.
          schema:
            $ref: '#/definitions/web.HTTPError'
        "500":
          description: Internal server error.
          schema:
            $ref: '#/definitions/models.Message'
      security:
      - JWTKeyAuth: []
      summary: Return user avatar setting
      tags:
      - user
    post:
      consumes:
      - application/json
      description: Changes the user avatar. Valid types are gravatar (uses the user
        email), upload, initials, marble, ldap (synced from LDAP server), openid (synced
        from OpenID provider), default.
      parameters:
      - description: The user's avatar setting
        in: body
        name: avatar
        required: true
        schema:
          $ref: '#/definitions/v1.UserAvatarProvider'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Message'
        "400":
          description: Something's invalid.
          schema:
            $ref: '#/definitions/web.HTTPError'
        "500":
          description: Internal server error.
          schema:
            $ref: '#/definitions/models.Message'
      security:
      - JWTKeyAuth: []
      summary: Set the user's avatar
      tags:
      - user
...
  /user/settings/general:
    post:
      consumes:
      - application/json
      parameters:
      - description: The updated user settings
        in: body
        name: avatar
        required: true
        schema:
          $ref: '#/definitions/models.UserGeneralSettings'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Message'
        "400":
          description: Something's invalid.
          schema:
            $ref: '#/definitions/web.HTTPError'
        "500":
          description: Internal server error.
          schema:
            $ref: '#/definitions/models.Message'
      security:
      - JWTKeyAuth: []
      summary: Change general user settings of the current user.
      tags:
      - user
...
  /user/timezones:
    get:
      consumes:
      - application/json
      description: Because available time zones depend on the system Brazn Tasks is running
        on, this endpoint returns a project of all valid time zones this particular
        Brazn Tasks instance can handle. The project of time zones is not sorted, you
        should sort it on the client.
      produces:
      - application/json
      responses:
        "200":
          description: All available time zones.
          schema:
            items:
              type: string
            type: array
        "500":
          description: Internal server error.
          schema:
            $ref: '#/definitions/models.Message'
      security:
      - JWTKeyAuth: []
      summary: Get all available time zones on this Brazn Tasks instance
      tags:
      - user
...
securityDefinitions:
  BasicAuth:
    type: basic
  JWTKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"